+++ assets/javascripts/kitten/components/notifications/alert/index.js
@@ -1,11 +1,11 @@
 import React, { useEffect, useRef, useState } from 'react'
 import PropTypes from 'prop-types'
 import styled, { keyframes } from 'styled-components'
-import { CloseButton } from '../../components/buttons/close-button'
-import COLORS from '../../constants/colors-config'
-import TYPOGRAPHY from '../../constants/typography-config'
-import { ScreenConfig } from '../../constants/screen-config'
-import { pxToRem, stepToRem } from '../../helpers/utils/typography'
+import { CloseButton } from '../../../components/buttons/close-button'
+import COLORS from '../../../constants/colors-config'
+import TYPOGRAPHY from '../../../constants/typography-config'
+import { ScreenConfig } from '../../../constants/screen-config'
+import { pxToRem, stepToRem } from '../../../helpers/utils/typography'
 import classNames from 'classnames'
 
 const fadeOut = keyframes`