+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Breaking change: Remove `k-Alert` SCSS mixin.
+
 ## [3.0.0-beta.1] - 2020-02-08
 
 Breaking changes: